refactor: Split types
